mini project/backend/requirements.txt
annotated-types==0.7.0
anyio==4.13.0
click==8.3.2
colorama==0.4.6
fastapi==0.115.3
h11==0.16.0
idna==3.11
numpy>=1.24.0
pillow>=10.0.0
pycryptodome>=3.18.0
pydantic==2.12.5
pydantic_core==2.41.5
python-multipart==0.0.26
starlette==0.41.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.32.0
